More minor refinement

+++ lib/common/functions.sh
@@ -1521,12 +1521,12 @@ configureNFS() {
     elif [[ -f "/usr/etc/nfs.conf" && ! -d "/usr/etc/nfs.conf.d/" ]]; then
         mkdir /usr/etc/nfs.conf.d
     fi
-    if [[ ! -f "/etc/nfs.conf.d/fog-nfs.conf" ]]; then
+    if [[ -f "/etc/nfs.conf" && ! -f "/etc/nfs.conf.d/fog-nfs.conf" ]]; then
         cat > /etc/nfs.conf.d/fog-nfs.conf <<EOF
 [mountd]
 port=20048
 EOF
-    elif [[ ! -f "/usr/etc/nfs.conf.d/fog-nfs.conf" ]]; then
+    elif [[ -f "/usr/etc/nfs.conf" && ! -f "/usr/etc/nfs.conf.d/fog-nfs.conf" ]]; then
         cat > /usr/etc/nfs.conf.d/fog-nfs.conf <<EOF
 [mountd]
 port=20048
